#2816: restructure and reimplement converter modules
