+++ brain/converter/svr/__init__.py
@@ -1,0 +1,4 @@
+'''
+This file allows the containing directory to be considered a python package,
+consisting of python module(s).
+'''
